Store extended colors in v:colornames dict.
Problem solved: Remembering RGB colors is difficult.
Solution: Allow user to (re-)define new color names.
