+++ runtime/doc/message.txt
@@ -141,6 +141,11 @@ larger.  When it's more there probably is an endless loop.  Probably a
 The color name {name} is unknown.  See |gui-colors| for a list of colors that
 are available on most systems.
 
+							*1204*  >
+  Bad color string: {str}
+
+The provided color did not conform to the pattern #rrggbb
+
 							*E458*  >
   Cannot allocate colormap entry, some colors may be incorrect
